move core check and update message and readme
