+++ src/commands/test.js
@@ -4,8 +4,6 @@ const _ = require('lodash');
 const constants = require('../constants');
 const utils = require('../utils');
 const validate = require('./validate');
-const updateNotifier = require('update-notifier');
-const path = require('path');
 
 const test = context => {
   const extraEnv = {
@@ -55,21 +53,6 @@ const test = context => {
             context.line(stdout);
           }
         });
-    })
-    .then(() => {
-      // find a package.json for the app and notify on the core dep
-      // `zapier test` won't run if package.json isn't there, so if we get to here we're good
-      const requiredVersion = _.get(
-        require(path.resolve('./package.json')),
-        `dependencies.${constants.PLATFORM_PACKAGE}`
-      );
-
-      if (requiredVersion) {
-        updateNotifier({
-          pkg: { name: constants.PLATFORM_PACKAGE, version: requiredVersion },
-          updateCheckInterval: constants.UPDATE_NOTIFICATION_INTERVAL
-        }).notify({ isGlobal: false });
-      }
     });
 };
 test.argsSpec = [];
